% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{ridgereg}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{the formula}

\item{data}{the data}

\item{lambda}{the lambda}
}
\value{
obj
}
\description{
ridgereg
}
\author{
hamed and Omid
}
