% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_qr.R
\name{ridgereg_qr}
\alias{ridgereg_qr}
\title{ridgereg_qr}
\usage{
ridgereg_qr(formula, data, lambda)
}
\arguments{
\item{formula}{the formula}

\item{data}{the data}

\item{lambda}{the lambda}
}
\value{
obj
}
\description{
ridgereg by qr decomposition
}
\author{
Hamed and Omid
}
